+++ packages/network/lib/agent.ts
@@ -15,23 +15,27 @@ const CRLF = '\r\n'
 const statusCodeRe = /^HTTP\/1.[01] (\d*)/
 
 let baseCaOptions: CaOptions | undefined
-const getCaOptionsPromise = (): Promise<CaOptions> => {
-  return getCaOptions().then((options: CaOptions) => {
+const getCaOptionsAsync = async (): Promise<CaOptions> => {
+  try {
+    const options = await getCaOptions()
+
     baseCaOptions = options
 
     return options
-  }).catch(() => {
+  } catch (error) {
+    debug('Error getting CA options', error)
+
     // Errors reading the config are treated as warnings by npm and node and handled by those processes separately
     // from what we're doing here.
-    return {}
-  })
+    return {} as CaOptions
+  }
 }
-let baseCaOptionsPromise: Promise<CaOptions> = getCaOptionsPromise()
+let baseCaOptionsPromise: Promise<CaOptions> = getCaOptionsAsync()
 
 // This is for testing purposes only
 export const _resetBaseCaOptionsPromise = () => {
   baseCaOptions = undefined
-  baseCaOptionsPromise = getCaOptionsPromise()
+  baseCaOptionsPromise = getCaOptionsAsync()
 }
 
 const mergeCAOptions = (options: https.RequestOptions, caOptions: CaOptions): https.RequestOptions => {
@@ -146,6 +150,7 @@ export const regenerateRequestHead = (req: http.ClientRequest) => {
   }
 }
 
+// this function has to be sync via callback because it is called by the Agent.addRequest method, which expect a sync function
 export const getFirstWorkingFamily = (
   { port, host }: Pick<http.RequestOptions, 'port' | 'host'>,
   familyCache: FamilyCache,
@@ -196,6 +201,7 @@ export class CombinedAgent {
   }
 
   // called by Node.js whenever a new request is made internally
+  // NOTE: this function has to be sync via callback because it is called by the Agent.addRequest method, which expect a sync function
   addRequest (req: http.ClientRequest, options: http.RequestOptions, port?: number, localAddress?: string) {
     _.merge(req, lenientOptions)
 
@@ -262,7 +268,7 @@ export class CombinedAgent {
   }
 }
 
-const getProxyOrTargetOverrideForUrl = (href) => {
+const getProxyOrTargetOverrideForUrl = (href: string) => {
   // HTTP_PROXY_TARGET_FOR_ORIGIN_REQUESTS is used for Cypress in Cypress E2E testing and will
   // force the parent Cypress server to treat the child Cypress server like a proxy without
   // having HTTP_PROXY set and will force traffic ONLY bound to that origin to behave
@@ -395,7 +401,7 @@ class HttpsAgent extends https.Agent {
     const port = options.uri?.port || '443'
     const hostname = options.uri?.hostname || 'localhost'
 
-    createProxySock({ proxy, shouldRetry: options.shouldRetry }, (originalErr?, proxySocket?, triggerRetry?) => {
+    createProxySock({ proxy, shouldRetry: options.shouldRetry }, (originalErr?: Error, proxySocket?: net.Socket, triggerRetry?: (err: Error) => void) => {
       if (originalErr) {
         const err: any = new Error(`A connection to the upstream proxy could not be established: ${originalErr.message}`)
 
@@ -406,12 +412,12 @@ class HttpsAgent extends https.Agent {
       }
 
       const onClose = () => {
-        triggerRetry(new Error('ERR_EMPTY_RESPONSE: The upstream proxy closed the socket after connecting but before sending a response.'))
+        triggerRetry?.(new Error('ERR_EMPTY_RESPONSE: The upstream proxy closed the socket after connecting but before sending a response.'))
       }
 
       const onError = (err: Error) => {
-        triggerRetry(err)
-        proxySocket.destroy()
+        triggerRetry?.(err)
+        proxySocket?.destroy()
       }
 
       let buffer = ''
@@ -423,15 +429,15 @@ class HttpsAgent extends https.Agent {
 
         if (!_.includes(buffer, _.repeat(CRLF, 2))) {
           // haven't received end of headers yet, keep buffering
-          proxySocket.once('data', onData)
+          proxySocket?.once('data', onData)
 
           return
         }
 
         // we've now gotten enough of a response not to retry
         // connecting to the proxy
-        proxySocket.removeListener('error', onError)
-        proxySocket.removeListener('close', onClose)
+        proxySocket?.removeListener('error', onError)
+        proxySocket?.removeListener('close', onClose)
 
         if (!isResponseStatusCode200(buffer)) {
           return cb(new Error(`Error establishing proxy connection. Response from server was: ${buffer}`), undefined)
@@ -453,14 +459,14 @@ class HttpsAgent extends https.Agent {
         cb(undefined, proxySocket)
       }
 
-      proxySocket.once('close', onClose)
-      proxySocket.once('error', onError)
-      proxySocket.once('data', onData)
+      proxySocket?.once('close', onClose)
+      proxySocket?.once('error', onError)
+      proxySocket?.once('data', onData)
 
       const connectReq = buildConnectReqHead(hostname, port, proxy)
 
-      proxySocket.setNoDelay(true)
-      proxySocket.write(connectReq)
+      proxySocket?.setNoDelay(true)
+      proxySocket?.write(connectReq)
     })
   }
 }